+++ packages/frontend-tk-cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aptx/frontend-tk-cli
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @aptx/frontend-tk-binding@0.1.4
+  - @aptx/frontend-tk-core@0.1.4
+  - @aptx/frontend-tk-plugin-aptx@0.1.4
+  - @aptx/frontend-tk-plugin-materal@0.1.4
+  - @aptx/frontend-tk-plugin-model@0.1.4
+  - @aptx/frontend-tk-plugin-input@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes